android: Fix misc compiler errors for GCC 5.4+
I'm preparing an update of our custom toolchains (trying both
with 5.4 and 6.2 at the moment), and using them reveals a lot
of minor compilation bugs:

- Most of them come from the fact that the compilers is _much_
  stricter when detecting 'signed vs unsigned' comparisons,
  and 'int vs size_t' ones.

  Fix the source code, including a lot of unit-tests where
  EXPECT_EQ() is used to compare ints and size_t values.

- Add some missing system headers (e.g. <ctype.h>) or
  C++ standard library ones (e.g. <memory> or <algorithm>)
  to some sources or headers.

- android/qemu1/util/aes.c: Move static variable declaration into
  conditional code block to avoid error about an unused variable.

- android/qemu1/include/qemu/osdep.h: Avoid redefining 'inline'
  as a macro when included from a C++ source file.

+++ android/android-emu/android/base/system/Win32UnicodeString_unittest.cpp
@@ -95,13 +95,13 @@ TEST(Win32UnicodeString, convertToUtf8) {
         char out[256];
         int len = Win32UnicodeString::convertToUtf8(out, sizeof(out),
                                                     kData[n].utf16);
-        EXPECT_EQ(strlen(kData[n].utf8) + 1, len);
+        EXPECT_EQ(strlen(kData[n].utf8) + 1U, (size_t)len);
         EXPECT_STREQ(kData[n].utf8, out);
 
         len = Win32UnicodeString::convertToUtf8(out, sizeof(out),
                                                 kData[n].utf16,
                                                 wcslen(kData[n].utf16));
-        EXPECT_EQ(strlen(kData[n].utf8), len);
+        EXPECT_EQ((int)strlen(kData[n].utf8), len);
         out[len] = 0;
         EXPECT_STREQ(kData[n].utf8, out);
 
@@ -121,11 +121,11 @@ TEST(Win32UnicodeString, convertToUtf8) {
         EXPECT_EQ(-1, len);
 
         len = Win32UnicodeString::calcUtf8BufferLength(kData[n].utf16);
-        EXPECT_EQ(strlen(kData[n].utf8) + 1, len);
+        EXPECT_EQ((int)strlen(kData[n].utf8) + 1, len);
 
         len = Win32UnicodeString::calcUtf8BufferLength(kData[n].utf16,
                                                       wcslen(kData[n].utf16));
-        EXPECT_EQ(strlen(kData[n].utf8), len);
+        EXPECT_EQ((int)strlen(kData[n].utf8), len);
     }
 }
 
@@ -145,13 +145,13 @@ TEST(Win32UnicodeString, convertFromUtf8) {
         wchar_t out[256];
         int len = Win32UnicodeString::convertFromUtf8(out, ARRAY_SIZE(out),
                                                       kData[n].utf8);
-        EXPECT_EQ(wcslen(kData[n].utf16) + 1, len);
+        EXPECT_EQ((int)wcslen(kData[n].utf16) + 1, len);
         EXPECT_STREQ(kData[n].utf16, out);
 
         len = Win32UnicodeString::convertFromUtf8(out, ARRAY_SIZE(out),
                                                   kData[n].utf8,
                                                   strlen(kData[n].utf8));
-        EXPECT_EQ(wcslen(kData[n].utf16), len);
+        EXPECT_EQ((int)wcslen(kData[n].utf16), len);
         out[len] = 0;
         EXPECT_STREQ(kData[n].utf16, out);
 
@@ -169,11 +169,11 @@ TEST(Win32UnicodeString, convertFromUtf8) {
         EXPECT_EQ(-1, len);
 
         len = Win32UnicodeString::calcUtf16BufferLength(kData[n].utf8);
-        EXPECT_EQ(wcslen(kData[n].utf16) + 1, len);
+        EXPECT_EQ((int)wcslen(kData[n].utf16) + 1, len);
 
         len = Win32UnicodeString::calcUtf16BufferLength(kData[n].utf8,
                                                        strlen(kData[n].utf8));
-        EXPECT_EQ(wcslen(kData[n].utf16), len);
+        EXPECT_EQ((int)wcslen(kData[n].utf16), len);
     }
 }
